Update debian packaging files for Ubuntu Focal
The i-nex package as configured fails to build on Ubuntu 20.04 LTS
"Focal Fossa" and does not build "lintian clean" on older versions.
This commit updates the entirety of the Debian package configuration and
brings it into compliance with the current Debian Standards Version,
v4.5.0.2. The debhelper compatibility level is bumped to 12, trading the
ability to build on Xenial for a package that should build successfully
for no less than the next four Ubuntu releases and Debian Bullseye.

Signed-off-by: Peter J. Mello <admin@petermello.net>